+++ example/axisnow/decryptAndEmbed.ts
