+++ example/axisnow/worker.ts
@@ -1,5 +1,5 @@
 import * as Comlink from 'comlink';
-import createChapterDecryptor from './decryptor';
+import createChapterDecryptor from './decryptAndEmbed';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 declare const self: WorkerGlobalScope;
